refactor: modularize backend and env config
