+++ carhub-api/utils/firebase.js
@@ -1,0 +1,3 @@
+const { Firestore, FieldValue, Timestamp } = require('@google-cloud/firestore');
+const firestore = new Firestore();
+module.exports = { firestore, FieldValue, Timestamp };
