+++ tsconfig.json
@@ -22,6 +22,7 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
     "noUncheckedSideEffectImports": true,
+    "types": ["vite/client"],
 
     "paths": {
       "@/*" :  ["./*"]
